+++ R/00-Namespace.R
@@ -2,9 +2,10 @@
 #' @rawNamespace import(DiagrammeR)
 
 #' @importFrom DataTools downloadModelServer downloadModelUI extractNotes extractObjectFromFile
-#'  importDataUI importDataServer importOptions
+#'  importOptions importServer importUI
 #' @importFrom magrittr "%>%"
-#' @importFrom jsonlite write_json fromJSON
+#' @importFrom openssl signature_create signature_verify base64_encode base64_decode
+#' @importFrom jsonlite write_json fromJSON toJSON
 #' @importFrom shinyalert shinyalert
 #' @importFrom yaml read_yaml