+++ inst/app/ui.R
@@ -18,7 +18,11 @@ tagList(
           # can be removed whe we have more inputs that can be stored in a session download:
           textInput("test_input", "Test Input", placeholder = "Please add some text ..."),
           tags$br(),
-          downloadModuleUI("download_unsigned"),
+          # this is only a placeholder for actual inputs,
+          # can be removed whe we have more inputs that can be stored in a session download:
+          textInput("test_input", "Test Input", placeholder = "Please add some text ..."),
+          tags$br(),
+          uiOutput("conditional_download_buttons"),
           checkboxInput(
             inputId = "download_inputs",
             label = "Download user inputs and graph",