+++ NEWS.md
@@ -1,6 +1,27 @@
+# TraceR 24.11.0
+
+## Bug Fixes
+
+- fix issue with zoom (#17)
+
+# TraceR 24.09.0
+
+## New Features
+
+- check enviroment for download/import enviroment
+
+# TraceR 24.09.0
+
+## New Features
+
+- option to sign the graph json file with a key pair
+- option to download & import signed json
+- deployment adjustments for shinyproxy
+
 # TraceR 24.09.0
 
 ## Bug Fixes
+
 - fix issue with zoom (#17)
 - add logo as icon for the browser tab